requirements.txt
requests
matplotlib
python-magic
elasticsearch